# services/api/cmd/showbiz-api/openapi.yaml
openapi: 3.0.3
info:
  title: Showbiz API
  description: Showbiz platform API service.
  version: 0.1.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /v1/organizations:
    post:
      operationId: createOrganization
      tags: [Organizations]
      summary: Create an organization
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                displayName:
                  type: string
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listOrganizations
      tags: [Organizations]
      summary: List organizations
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Paginated list of organizations
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedList"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Organization"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/organizations/{id}/members:
    get:
      operationId: listMembers
      tags: [Members]
      summary: List organization members
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/organizationId"
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: addMember
      tags: [Members]
      summary: Add a member to an organization
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/organizationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "204":
          description: Member added
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/organizations/{id}/members/{email}:
    delete:
      operationId: removeMember
      tags: [Members]
      summary: Remove a member from an organization
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/organizationId"
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
      responses:
        "204":
          description: Member removed
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Organization or member not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Projects ────────────────────────────────────────────────────────
...
  /v1/iam/policies/{policyId}:
    get:
      operationId: getPolicy
      tags: [IAM Policies]
      summary: Get a policy by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/policyId"
      responses:
        "200":
          description: Policy details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Policy not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Reusable Parameters ─────────────────────────────────────────────
  parameters:
...
    organizationId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    policyId:
      name: policyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: Pagination cursor returned by a previous request.
    limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Maximum number of items to return.

  # ── Schemas ─────────────────────────────────────────────────────────
  schemas:
    User:
      type: object
      properties:
        email:
          type: string
          format: email
        organizationId:
          type: string
        displayName:
          type: string
        emailVerified:
          type: boolean
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Organization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        displayName:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Policy:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        scope:
          type: string
          enum: [global, organization]
        organizationId:
          type: string
          format: uuid
          nullable: true
        permissions:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
    PaginatedList:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items: {}
        pagination:
          type: object
          required: [nextCursor, hasMore]
          properties:
            nextCursor:
              type: string
              nullable: true
            hasMore:
              type: boolean